+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Select Colony, MS Pipeline, and Bid points on focus (making it easier
   for mobile users to update the values)
+- Space wreck Randomizer added to CP tab.
 
 ## 3.1.0